chore: move assetsEntry type to shared
